// lib/youtube/api.ts
export interface YouTubeChannel {
  id: string;
  title: string;
  handle: string;
  description: string;
  subscriberCount: number;
  videoCount: number;
  viewCount: number;
  thumbnailUrl: string;
  bannerUrl: string;
}